Initial version
For now beaver is basically one meta-command that has sub-commands and
extensions to perform common tasks. In this initial version there is
a docker extension that has 2 commands: build and run. This is intended
to standardize how docker images are build and run in beaver projects.

There is also a bintray extension that for now has only an upload
command to upload Debian packages, since Travis build-in support is not
the most convenient.

Then there is the install beaver sub-command that just calls docker
build in a more specific way, that should be more convenient for most
projects. This command can perform other setup steps in the future that
are not related to docker.

A draft README file is also provided, but for now it acts more as
a reminder of how things work than proper end-user documentation.