Add auto-discovered single tests
It doesn't scale well to run all the tests right from the .travis.yml file.
Is good to test the basic layout of a .travis.yml for a beaver project but
beyond that, tests can live separately and just ran once in Travis.

This adds a new matric cell to just run these tests, separately from the
travis matrix tests.