gen-dockerfile: Fix docker repository and tag parsing
It is quite hard to find a good reference on which characters are valid
for each part of the docker repository and tag.

The user name seems to be restricted to [0-9a-z] from experience
creating users at Docker Hub.

"The Repository Name will need to be unique in that namespace, can be
two to 255 characters, and can only contain lowercase letters, numbers
or - and _." according to
https://docs.docker.com/docker-hub/repos/#creating-a-new-repository-on-docker-hub.

"A tag name must be valid ASCII and may contain lowercase and uppercase
letters, digits, underscores, periods and dashes. A tag name may not
start with a period or a dash and may contain a maximum of 128
characters." according to
https://docs.docker.com/engine/reference/commandline/tag/.

Nevertheless, to try to be as future-proof as possible, we are trying to
parse just based on the field separators (spaces, "/", ":" and "#").

Add some more testing for the FROM line conversion too.