beaver install: Support multiple Dockerfile contexts
Sometimes is useful to have multiple different docker contexts for
a project. The most typical case is having one (`beaver.`)`Dockerfile`
for building a project (with the build dependencies), and another one to
run the project (with the runtime dependencies). Since Docker doesn't
support specifying a `.dockerignore` file yet, we overcome this
limitation in beaver.

To use multiple build contexts, you can specify the variable
`BEAVER_DOCKER_CONTEXT` to point to a directory that will contain your
`Dockerfile` or `beaver.Dockerfile` and possibly a `build` script
(equivalent to the `docker/build` script in normal builds) and/or
a `dockerignore` (`.dockerignore is also accepted but `dockerignore`
will take precedence), which will be copied to the current working
directory as `.dockerignore`.